pyproject.toml
# https://setuptools.pypa.io/en/latest/userguide/pyproject_config.html

[build-system]
build-backend = "flit_core.buildapi"
requires = ["flit_core >=3.2,<4"]

[project]
authors = [
  {name = "UT", email = "joyce.poon@utoronto.ca"}
]
classifiers = [
  #"Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Operating System :: OS Independent"
]
dependencies = [
    "gdsfactory==9.8.0",
    "gplugins[sax,tidy3d]==1.3.8",
    "tidy3d==2.8.4",
    "sax==0.13.3",
    "fastapi",
    "uvicorn",
    "sse-starlette",
    "sseclient",
    "pyyaml",
    "numpy",
    "pandas",
    "rank_bm25",
    "streamlit",
    "jupyter",
    "ipykernel",
    "torch",
    "torchvision",
    "torchaudio",
    "transformers[torch]",
    "scikit-optimize",
    "groq",
    "openai==1.43.0",
    "fireworks-ai",
    "replicate",
    "sentence_transformers",
    "nltk",
    "pyswarms",
    "google-generativeai",
    "mpld3",
    "pygraphviz",
    "tiktoken",
    "haystack-ai",
    "bayesian-optimization",
    "protobuf==3.20.3",
]
description = "Design"
keywords = ["python"]
license = {file = "LICENSE"}
name = "PhotonicsAI"
readme = "README.md"
requires-python = ">=3.11"
version = "0.0.1"

[project.optional-dependencies]
dev = [
  "pre-commit",
  "pytest",
  "pytest-cov",
  "pytest_regressions"
]
docs = [
  "jupytext",
  "matplotlib",
  "jupyter-book>=0.15.1,<1.1"
]

[tool.codespell]
ignore-words-list = "te, te/tm, te, ba, fpr, fpr_spacing, ro, nd, donot, schem"

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pylsp-mypy]
enabled = true
live_mode = true
strict = true

[tool.pytest.ini_options]
# addopts = --tb=no
addopts = '--tb=short'
norecursedirs = ["extra/*.py"]
python_files = ["PhotonicsAI/*.py", "notebooks/*.ipynb", "tests/*.py"]
testpaths = ["PhotonicsAI/", "tests"]

[tool.ruff]
exclude = [
  "deprecated/*"
]
fix = true
ignore = [
  "E501",  # line too long, handled by black
  "B008",  # do not perform function calls in argument defaults
  "C901",  # too complex
  "B905",  # `zip()` without an explicit `strict=` parameter
  "C408"  # C408 Unnecessary `dict` call (rewrite as a literal)
]
select = [
  "B",  # flake8-bugbear
  "C",  # flake8-comprehensions
  # "D",  # pydocstyle
  "E",  # pycodestyle errors
  "F",  # pyflakes
  "I",  # isort
  "T10",  # flake8-debugger
  "UP",  # pyupgrade
  "W"  # pycodestyle warnings
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.setuptools.package-data]
mypkg = ["*.csv", "*.yaml"]

[tool.setuptools.packages]
find = {}

[tool.tbump]

[[tool.tbump.file]]
src = "README.md"

[[tool.tbump.file]]
src = "pyproject.toml"

[[tool.tbump.file]]
src = "PhotonicsAI/__init__.py"

[tool.tbump.git]
message_template = "Bump to {new_version}"
tag_template = "v{new_version}"

[tool.tbump.version]
current = "0.0.1"
# Example of a semver regexp.
# Make sure this matches current_version before
# using tbump
regex = '''
  (?P<major>\d+)
  \.
  (?P<minor>\d+)
  \.
  (?P<patch>\d+)
  '''